feat: Add non-standard helpers for checking args and blocks
We need to ensure users have either a @label or :label block. Any sort of label is required.

We also need to check if users have either an @arg or a block to render a component.

Because we are using glint, we currently can't use normal helpers. So we have -private ones instead.